+++ mutraction-dom/src/index.ts
@@ -3,7 +3,6 @@ export { ForEach, ForEachPersist } from './foreach.js';
 export { choose } from './choose.js';
 export { isTracked } from './proxy.js';
 export { PromiseLoader } from './promiseLoader.js';
-export { ErrorBoundary } from './errorBoundary.js';
 export { Swapper } from './swapper.js';
 export { Tracker, TrackerOptions, defaultTracker, track } from './tracker.js';
 export { PropReference, createOrRetrievePropRef } from './propref.js';