+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- `ErrorBoundary` - turns out it wasn't that good
+
 ## [0.19.1] - 2023-09-05
 
 ### Changed